requirements.txt
# Python Dependencies for GBM Classification Project
# Hogeschool Rotterdam - Minor AI in Healthcare
# Last Updated: November 2025

# Core Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0

# Batch Correction
combat>=0.3.3
pycombat>=0.3.0
# harmonypy>=0.0.9  # Uncomment if using Harmony

# Dimensionality Reduction
umap-learn>=0.5.0

# Model Explainability
shap>=0.43.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration & I/O
pyyaml>=6.0
joblib>=1.3.0

# Progress Bars (optional but useful)
tqdm>=4.65.0

# Jupyter Notebook Support (for EDA notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.0.0

# Code Quality (development - optional)
# black>=23.0.0
# flake8>=6.0.0
# pytest>=7.4.0